Removed the Barbarian Unarmored defense selection in favor of an automatic  bonus to AC when given the ability that is removed when the player equips heavy armor.
Fixed a bug with adding Gnome Tinker's tools ability.

Fixed define for Slight of hand skill to match the bonus given.

Added bonus attack that should show on character sheet with the barbarian extra attack ability.

Added bonus to stats for barbarian primal champion ability.

Changed barbarian path of the totem warrior to give abilities associated with path progress.

Added ritual only versions of the spells Beast sense, Speak with animals and Commune with Nature.

Path of the totem warrior gives these spells automatically with path progress.

Changed totem spirit abilities to separate spirit animal choice into separate displayed abilities.

Aspect of the bear now gives bonus to load multipliers. I had to change the load.lst in order to do this correctly for small and smaller creatures due to the fact that you can't give fractional bonuses. The base weight loads are now adjusted so that a fine creature's SIZEMULT is 1. The load for each point of strength is adjusted to be 5/8 of what it was. The loads for each creature should be the same but this will need to be further verified with things such as item weight. The line that gives these bonuses will have to be changed if/when fine and diminutive size categories become available.

Bard Spellcasting: I have changed the way spell replacement works by simply using the Known: tag in the classes.lst file. I calculate the number of known spells operating under the assumption that for that level; one of the lowest spell slot levels will be replaced by the highest, as well as adding new known spells to the highest. For the bard, these numbers are adjusted to account for the fact that the magical secrets ability counts for the new spells learned during level progress (but not when it's part of the college progress line). The effect should be that with the fact that PCGen will allow a lower level spell to take the place of an available higher level one, any spell switching should work as the rules describe.

Bardic inspiration should now display the exact dice used as well as the exact number of times available to use.

Jack of all trades should give the rounded down bonus to all skills that the character is not proficient in.

Song of rest should now display the exact dice used.

Magical Secrets and Additional Magical secrets should now give the character the ability to select spells from all the spell lists limited by the level the character is able to cast. These will automatically be added to the Bard spell list.

Added bonus attack that should show on character sheet with the bard extra attack ability.