Unapproachable East Class Mod
So myrlyn made a commit on 20160307 that added a feat, Craft Nar
Demonbinder Sign, to match the VFEAT I left in the class abilities file.
Initially the VFEAT entry was supposed to have been removed, this was a
mistake on my part.

In my own campaigns I allow significant class features to act as feats
of various categories which allows for interaction with other feats and
prereqs. This wasn't supposed to make it into the main branch since it
is not strictly a rule (for lack of any statement within the description
indicating this). It's something that requires adjudication.

To accommodate for this, I created two templates that would VFEAT the
two new feats I created (one of which replaces myrlyn's Create Nar
Demonbinder Sign). I specifically moved this functionality into
templates and out of the class and class ability files so that nothing
about the functionality of the class required it. The templates could be
applied based upon the house rules of the individual campaign without
threatening the functionality of the class as implemented following the
rule books.