Unapproachable East Class Mod
Replaced the VFEAT tag with the ABILITY tag in the templates file.
Changed the VISIBLE tag from NO to EXPORT in the feats file. Left the
PRECLASS tag to control access when other selectors pop the ItemCreation
feats. Since these abilities are restricted to specific classes but
might, because of their type, be pulled into other lists, their
availability needs to remain nailed down.
Replaced the CLASSTYPE tag with the FACT:CLASSTYPE tag in the class
file.