NarDemonbinderUpdates (#676)
* NarDemonbinderUpdates

Corrected a typo. Made a vain effort to restrict the class' spell list,
most specifically against things added elsewhere with a CLASSES=ALL
tagging.

* NarDmonbinderUpdates

Corrected a typo. Made a vain effort to restrict the class' spell list,
most specifically against things added elsewhere with a CLASSES=ALL
tagging. Removed some redundancy in the spells.lst between the
declaration of the spell entry and the .MOD entries at the end of the
file.