DraconicHeritageFeatCorrection
Corrected the pre-reqs for the feats, which in the case of the listings
in the Complete Arcane made it impossible to take any of them as it was
dependent upon a variable that was never defined.
Made the adjustments to the versions in the Races of the Dragon even
though some aspects of them were redundant.