Further mworka and uncanny dodge fixes (#832)
* Update ph_deities.lst

Updated Fharlanghn entry as was missing a domain: Travel.

* Duskblade bug fixes

Duskblades are spontaneous casters and spells known were not being enforced correctly, but are now. Minor issue with the armourmed mage ability description also fixed.

* Further MWORKA and Uncanny Dodge updates

Updating of MWORKA to Masterwork (Armor) and Uncanny Dodge to Uncanny Dodge ~ Base in further books.