C bo em for 688 with extra desc (#1105)
* Update sc_spells.lst

updated spell descriptions and renamed one spell to avoid conflict with spell from PH2

* Night Haunt - typo on desc

fixed a typo

* CBoEM for 688 with desc

repeat the previous commit for CBoEM but this time for 688, and add some additional feat descriptions.

Bahamut - please confirm - I'm only using 688 (due to issues with running v9) so should I commit to 688 branch only and get you to 'import' to v9 as well - or vice versa? What is the best approach to make updates for both?