Fixed "Brains over Brawn" ability of Factotum
The ability gives the intelligence modifier as a bonus, instead of replacing the original ability modifier.

While the ability description might seem to indicate that it is a replacement, it does not state so.
This is supported by the skills stat block from the 6th-level example NPC "Alerach Longseeker" in the book; those skill levels can only be obtained if the Brains over Brawn ability works that way:
- A +14 on Open Lock corresponds to 9 ranks (maximum number of ranks for a 6th-level character), +2 from Dex, and +3 from Int.
- None of the Str/Dex-based skills have values below the Dex/Str modifier + Int modifier, and Alerach does not have enough skill points to get these skills to that level if the Brains over Brawn ability would replace instead of add on top of the existing ability modifier.
(I consider the "Search +13" in the stat block a typo, it should be +12 instead.)