Fixed Bonus Inspiration point calculation
Currently the implementation grants one inspiration point per "Font of Inspiration" feat, which is incorrect. The first feat grants 1, the second grants 2, the third 3, etc. With this change the bonus is calculated as the sum of all integers from 1 to the number of "Font of Inspiration" feats taken ((n*(n+1)/2).