Fix Warlock class' Eldritch Blast damage
The formula for EB damage was adding 1/2 to the warlock level instead of multiplying by 1/2.  This change makes the damage for characters up to level 18 consistent with Table 1-1.  The pattern up to this point would give the final d6 of damage at level 19 (which is what this change does), but the book gives it at level 20 for some reason.